Moved the canvas drawing functions to the functions file.
