+++ client/src/main/scripts/components/ShareContainer.tsx
@@ -3,6 +3,11 @@ import * as ReactDOM from "react-dom";
 import { TakeDocument, DocumentBlock, VideoBlock } from "./BlockEditor";
 import {
   alertErr,
+  drawCaption,
+  drawDocument,
+  drawDocumentText,
+  drawSpecs,
+  drawText,
   getCaptionNodeArray,
   getCharRangeFromVideoRange,
   getHighlightedNodes,
@@ -28,20 +33,10 @@ interface VideoImageURIs {
   captions: string | null;
 }
 
-interface Dimensions {
-  x: number;
-  y: number;
-  totalHeight: number;
-}
-
 class ShareContainer extends React.Component<
   ShareContainerProps,
   ShareContainerState
 > {
-  private textMargin: number = 16;
-  private width = 500;
-  private linewidth = 468;
-  private lineheight = 1.5; //multiplier
   constructor(props: ShareContainerProps) {
     super(props);
 
@@ -51,220 +46,6 @@ class ShareContainer extends React.Component<
       }
     };
   }
-  drawDocumentText = (
-    ctx: CanvasRenderingContext2D,
-    nodes: FoundationNode[],
-    title: string
-  ): number => {
-    // Draw fact title
-    const titleSize = 20;
-    let textSize = titleSize;
-    ctx.font = "Bold " + textSize.toString() + "px Source Sans Pro";
-    let x = this.textMargin;
-    let y = textSize;
-    ctx.fillText(title, x, y);
-
-    for (const node of nodes) {
-      if (node.component === "h2") {
-        // Set the font style
-        textSize = 22.5;
-        ctx.font = "Bold " + textSize.toString() + "px Merriweather";
-
-        // Add a margin above the new line of text
-        y += textSize * this.lineheight;
-
-        // Initialize an empty line of texxt
-        let line = "";
-        for (const text of node.innerHTML) {
-          // Loop through the innerHTML array to search for React Elements
-          if (text) {
-            let textStr = text.toString();
-            if (textStr === "[object Object]") {
-              // Can't find a better conditional test
-              // Found a React Element
-              line += (text as ReactElement<HTMLSpanElement>).props.children;
-            } else {
-              line += textStr;
-            }
-          }
-        }
-
-        // Write the line of text at the coordinates
-        ctx.fillText(line, this.textMargin, y);
-      } else if (node.component === "p") {
-        // Set font size
-        textSize = 15;
-
-        // Initialize the coorindates
-        x = this.textMargin; // Left margin of text
-        y += textSize * this.lineheight * this.lineheight; // Top margin
-
-        // Loop through the innerHTML array to search for React Elements
-        for (const text of node.innerHTML) {
-          if (text) {
-            let textStr = text.toString();
-            let words = "";
-            if (textStr === "[object Object]") {
-              // Can't find a better conditional test
-              // Found a React Element
-              words += (text as ReactElement<HTMLSpanElement>).props.children;
-              // This text is highlighted, so make it bold
-              ctx.font = "Bold " + textSize.toString() + "px Merriweather";
-            } else {
-              words += textStr.trim();
-              ctx.font = textSize.toString() + "px Merriweather";
-            }
-
-            const dimensions = this.drawText(ctx, words, textSize, x, y);
-            // Set the dimensions of the next line to be drawn where the previous line left off
-            y = dimensions.y;
-            x = this.textMargin + dimensions.x;
-          }
-        }
-        y += textSize / 2; //bottom margin
-      } else {
-        const errStr = "Unknown component";
-        alertErr(errStr);
-        throw errStr;
-      }
-    }
-    return y;
-  };
-  drawDocument = (nodes: FoundationNode[], title: string): string => {
-    const canvas = document.createElement("canvas");
-    const ctx = canvas.getContext("2d");
-
-    canvas.width = this.width * window.devicePixelRatio;
-    canvas.style.width = this.width + "px";
-
-    if (ctx) {
-      // Draw the document once to calculate height
-      const height = this.drawDocumentText(ctx, [...nodes], title);
-
-      canvas.height = height * window.devicePixelRatio;
-      canvas.style.height = height + "px";
-
-      ctx.scale(window.devicePixelRatio, window.devicePixelRatio);
-
-      // Draw grey background
-      ctx.fillStyle = "#f2f4f7";
-      ctx.fillRect(0, 0, this.width, height);
-
-      // Set text color
-      ctx.fillStyle = "#051a38";
-
-      // Draw document again to draw the text
-      this.drawDocumentText(ctx, [...nodes], title);
-
-      return canvas.toDataURL("image/png");
-    } else {
-      const errStr = "Error getting canvas context";
-      alertErr(errStr);
-      throw errStr;
-    }
-  };
-  drawCaption = (text: string): string => {
-    const canvas = document.createElement("canvas");
-    const ctx = canvas.getContext("2d");
-
-    canvas.width = this.width * window.devicePixelRatio;
-    canvas.style.width = this.width + "px";
-
-    if (ctx) {
-      // Set font styles
-      const textSize = 15;
-      ctx.font = "Bold " + textSize.toString() + "px Merriweather";
-
-      // Draw text once to calculate height
-      const height = this.drawText(ctx, text, textSize).totalHeight;
-
-      canvas.height = height * window.devicePixelRatio;
-      canvas.style.height = height + "px";
-
-      ctx.scale(window.devicePixelRatio, window.devicePixelRatio);
-
-      // Draw grey background
-      ctx.fillStyle = "#f2f4f7";
-      ctx.fillRect(0, 0, this.width, height);
-      ctx.fillStyle = "#051a38";
-
-      // Not sure why, but font has been reset at this point, so must set it again
-      ctx.font = "Bold " + textSize.toString() + "px Merriweather";
-      this.drawText(ctx, text, textSize);
-
-      return canvas.toDataURL("image/png");
-    } else {
-      const errStr = "Error getting canvas context";
-      alertErr(errStr);
-      throw errStr;
-    }
-  };
-  drawText = (
-    ctx: CanvasRenderingContext2D,
-    words: string,
-    fontsize: number,
-    initialX?: number,
-    initialY?: number
-  ): Dimensions => {
-    let wordsArr = words.split(" ");
-    // Initialize variables
-    let currentLineWidth = 0;
-    let isFirstLine = true;
-    let line = "";
-    let x = initialX ? initialX : this.textMargin;
-    let y = initialY ? initialY : fontsize;
-    let totalHeight = fontsize;
-
-    for (const word of wordsArr) {
-      if (isFirstLine && initialX) {
-        // Need to include width of previous line in this case
-        currentLineWidth = ctx.measureText(line + word).width + initialX;
-      } else {
-        currentLineWidth = ctx.measureText(line + word).width;
-      }
-
-      if (currentLineWidth > this.linewidth) {
-        // Start a new line
-        if (isFirstLine) {
-          if (!initialX) {
-            y += fontsize / 2; //top margin of new paragraph
-          }
-          isFirstLine = false;
-        } else {
-          y += fontsize * this.lineheight;
-        }
-        ctx.fillText(line, x, y);
-        totalHeight += fontsize * this.lineheight;
-        x = this.textMargin;
-        line = word + " ";
-      } else {
-        line += word + " ";
-      }
-    }
-
-    // Draw the last line
-    if (line.length > 0) {
-      if (isFirstLine) {
-        if (!initialX) {
-          y += fontsize / 2; //top margin of new paragraph
-        }
-        isFirstLine = false;
-      } else {
-        y += fontsize * this.lineheight;
-      }
-      totalHeight += fontsize * this.lineheight;
-      ctx.fillText(line, x, y);
-      x = this.textMargin;
-    }
-
-    totalHeight += fontsize / 2; // add bottom margin
-
-    return {
-      x: ctx.measureText(line).width,
-      y: y,
-      totalHeight: totalHeight
-    };
-  };
   getFacts = (
     done: (
       error: string | Error | null,
@@ -335,7 +116,7 @@ class ShareContainer extends React.Component<
                   blockInScope.viewRange
                 );
 
-                const url = this.drawDocument(
+                const url = drawDocument(
                   [...highlightedNodes],
                   factContent.fact.title
                 );
@@ -413,7 +194,7 @@ class ShareContainer extends React.Component<
                   highlightedText = highlightedText.trimRight();
                   highlightedText += '"';
 
-                  const uri = this.drawCaption(highlightedText);
+                  const uri = drawCaption(highlightedText);
 
                   videoFacts[index] = {
                     youtube: factContent.youtubeId,
@@ -490,7 +271,11 @@ class ShareContainer extends React.Component<
             "</tr>";
 
           const imageStyles =
-            "width:" + this.width + "px;" + "height:auto;" + "max-width:100%;";
+            "width:" +
+            drawSpecs.width +
+            "px;" +
+            "height:auto;" +
+            "max-width:100%;";
 
           // Loop through blocks
           for (const block of this.props.takeDocument.blocks) {
